app/requirements.txt
APScheduler==3.7.0
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.1
Werkzeug==1.0.1
click==7.1.2
cycler==0.10.0
implicit==0.4.4
itsdangerous==1.1.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mlxtend==0.18.0
numpy==1.20.1
pairwise==0.1
pandas==1.3.4
pip==21.0.1
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
setuptools==54.0.0
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.58.0
tzlocal==2.1
gunicorn==20.1.0